requirements.txt
absl-py==2.0.0
addict==2.4.0
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
asgiref==3.7.2
AutoTorch==0.0.1
backports.zoneinfo==0.2.1
bcrypt==4.1.2
blinker==1.6.3
cachetools==5.3.1
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==3.0.0
cmake==3.25.0
colorama==0.4.6
coloredlogs==15.0.1
ConfigSpace==0.4.11
contourpy==1.1.1
crcmod==1.7
cryptography==41.0.7
cycler==0.12.1
Cython==0.29.33
dask==2023.5.0
DBUtils==1.3
distributed==2023.5.0
Django==4.2.7
easydict==1.10
faiss-gpu==1.7.2
filelock==3.9.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.47.0
fsspec==2023.12.2
google-auth==2.23.2
google-auth-oauthlib==1.0.0
graphviz==0.8.4
grpcio==1.59.0
humanfriendly==10.0
idna==3.4
imageio==2.31.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lit==15.0.7
locket==1.0.0
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.4
mdurl==0.1.2
mmcv==1.3.18
# Editable install with no version control (mmdet==2.7.0)
-e /home/cv/zhouhs/projects/insightface/detection/scrfd
mmengine==0.10.1
mmpycocotools==12.0.3
model-index==0.1.11
mpmath==1.3.0
msgpack==1.0.7
mxnet==1.9.1
networkx==3.0
nose==1.3.7
numpy==1.22.3
oauthlib==3.2.2
onnx==1.14.1
onnxruntime-gpu==1.15.0
opencv-python==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.32
ordered-set==4.1.0
oss2==2.17.0
packaging==23.2
pandas==2.0.3
paramiko==2.12.0
partd==1.4.1
Pillow==9.3.0
platformdirs==4.1.0
protobuf==4.24.4
psutil==5.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.19.0
Pygments==2.17.2
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.10.1
shapely==2.0.2
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.4
sympy==1.12
tabulate==0.9.0
tblib==3.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.2.0
tifffile==2023.7.10
tomli==2.0.1
toolz==0.12.0
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tornado==6.4
tqdm==4.65.2
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
Werkzeug==3.0.0
yapf==0.40.2
zict==3.0.0
zipp==3.17.0
